{util-stats,twitter-server}: Make it possible to configure counter-ish gauges
Problem

Some metrics we don't control are essentially counters, tracking the amount of
something since the start of an application.  However, we don't have a good way
of signalling that.

Solution

Although we considered changing them to be implemented as counters, we were
concerned that folks who use latched counters would be surprised that these
metrics suddenly started being latched.  Although we want to transition in that
direction eventually, we chose to do it more cautiously.  Instead, we kept
the status quo, but added a "counterish_gauge" flag to the metrics metadata,
so that if a machine is trying to parse it, they can infer that they may want to
take the difference between querying the metrics endpoint.

Result

The jvm cycles and msec metrics can be more correctly reported!

JIRA Issues: CSL-10752

Differential Revision: https://phabricator.twitter.biz/D629892