util-validator: Use org.scalacheck.Shrink.shrinkAny in tests
Problem

When using ScalaCheck in tests to generate failing values, shrinking
the failing values may generate invalid values as the constraints of
the generator being used are not respected.

Solution

In the relevant tests, import `org.scalacheck.Shrink.shrinkAny` to
change the behavior of `forAll` to disable shrinking.

Result

Less brittle code to deal with the invalid values being produced
by the generator when testing failure values.

Differential Revision: https://phabricator.twitter.biz/D661479