util-core finagle-core: introduces forking scheduler
Problem

The adaptive executor project is being developed in a separate module. One of the goals of the project is
deprecating the need to use separate `FuturePool`s for the execution of CPU intensive and blocking tasks,
allowing users to rely on the scheduler to handle the different types of tasks in a finagle application. It's
not possible to provide a seamless migration path to this new behavior without changing the `FuturePool`
implementation.

Solution

This review introduces a `ForkingScheduler` interface to be extended by the fiber scheduler
(implementation of the adaptive executor project) and introduces a mechanism for the scheduler
to define if the future pool calls should be redirected to the forking scheduler.

Notes

- I placed the new interface in util since I expect that util code might rely on it in the future
- There's no behavior change in case the application doesn't use a forking scheduler
- I fixed the misplaced comment about the tracing filter in the server stack

JIRA Issues: VM-4488

Differential Revision: https://phabricator.twitter.biz/D644985