util: Add util-jackson
Problem

We'd like a general purpose JSON parsing library which provides a
level of support over the Jackson and `jackson-module-scala` libraries.

Solution

The core of the Finatra Jackson support has proven to be useful in
providing utility for JSON serialization and deserialization. Thus we are
lifting the core of the Finatra Jackson library to be general purpose
JSON parsing with Jackson in Scala.

Result

A usable JSON parsing library with an enhanced Scala API and improved
case class deserializer for collecting deserialization failures.

Differential Revision: https://phabricator.twitter.biz/D664962