util-app: Fix RealCommandTest when run via sbt
Problem: RealCommandTest does not work when run via sbt because the current
working directory is different than when the test is run via pants.

Solution: Use the path to the resource in the classloader to find the test
script rather than making an assumption about where it is.

JIRA Issues: CSL-11042

Differential Revision: https://phabricator.twitter.biz/D694390