util-jackson: Include null values in JsonDiff#toSortedString
Problem

`JsonDiff#assertDiff` writes a report that is meant to show where JSON objects
differ. It shows each object using `toSortedString` so that key ordering is
insignificant in the report.  But `#toSortedString` currently filters null
values.  The result is that when objects differ only in extra keys with null
values, the `JsonDiff#assertDiff` report shows identical string representations
and fails to show where the objects differ.

Solution

Include null values in the `#toSortedString` representation.

Differential Revision: https://phabricator.twitter.biz/D707033