util/util-core: scala 3 migration
**Problem**

Build util-core with Scala 3

**Solution**

We first

1. util-core project settings.
2. Use Scala 3 migration flags
3. Fix incompatibilities

Note: We keep the migration flags for now until we fix all
warnings (mainly syntax) with the automatic compiler rewrite in
a follow up diff.

**Result**

Code changes were made in. The area's the code changes were made
in provide comments linking tickets and github issues.

1. **AsyncStream**

Remove `Cons.unapply` due to dotty type parameter
non-exhaustive pattern match with AsyncStream's covariant type
(inclined to think this is a dotty bug). `Cons.unapply` is not used in
source as `Cons` is an internal type. We can workaround this issue
by making the constructor private and defining our own `apply`.

2. **Promise**
`Promise#raise` is not a tailrec implementation. We make it tailrec
by defining a private method `raiseImpl` that the dotty compiler
recognizes as tail recursive

3. **Reader**

Seq#hasDefiniteSize deprecations leading to `SeqUtil`.

4. **Broker**

unreachable code when pattern matching on joined types (A | B). To
be both Scala2 and Scala3 compatible, we can simply compare equality
with the case object `Quiet` and split the cases, removing the need for
a combined pattern match

5. **ClosableOnce**

Java compilation error declaring `AbstractClosableOnce` was not overriding
`close` even though the `CloseOnce` mixin is apart of the `ClosableOnce`
trait. This is perhaps a dotty bug. For now, we can make the `CloseOnce`
trait extend `Closable` instead of self typing with `Closable`.

6. **Spool**

Remove `AbstractSpool`. `Spool` is a sealed trait in scala so it's not intended
to be extended elsewhere. `AbstractSpool` is not java compatible in dotty
due to type erasure and the overloaded `def ++` in Scala. Let's remove AbstractSpool
since it shouldn't be used anyways to create Spools.

7. **Duration/Event**

Redone pattern match to avoid compiler warnings around exhaustiveness

8. **StorageUnit**

Numeric's `signum` is deprecated in favor of `sig`. Instead of splitting code, the
implementation of `signum` is so simple, let's just define our own version until
scala 2.12 support is dropped. Added test to make sure that method is tested.

JIRA Issues: CSL-11082

Differential Revision: https://phabricator.twitter.biz/D694775