util/util-core: fix scala3 runtime problems
Problem:
Scala 3 will optimize away `private[this]` members
that are not used within the class. This breaks the
`DetachedPromise` implementation causing the tests
and runtime usages of Promise to fail.

Solution:
Fix this by adding a private method that accesses this
field so the compiler will not remove the field.

Result:
There's also a compiler bug in the 3.0.1 release when
pattern matching on private inner classes. Let's fix a test
that uses this pattern until the next Scala 3 release

JIRA Issues: CSL-11232

Differential Revision: https://phabricator.twitter.biz/D718074