util/util-thrift: scala3 migration
Problem

We're migrating util-codec to scala3. Dotty has new syntax rules that need to be
addressed.

Solution

Use migration flags to solve syntax warnings, in this case there was no need for
any syntax changes.  While Manifest has not been dropped in scala3, it does not
work the same way as it did in scala2, for this reason we are using ClassTag now

JIRA Issues: CSL-11108

Differential Revision: https://phabricator.twitter.biz/D715129