util: crossbuild with the Scala 3.0.2-RC1
Problem:

The Scala 3.0.1 compiler includes bugs that are patched
in the 3.0.2-RC1 release. An example of this is the broken
Java APIs in JDK 11 for our `TimeLikeOps` classes because of
https://github.com/lampepfl/dotty/issues/13039.

Result / Solution:
Using the 3.0.2-RC1 release doesn't break any tests and our
projects compile fine. Let's update to this to make use of the bug
fixes until the 3.0.2 is officially released.

The new Dotty scaladoc requires network access due to a bug
https://github.com/lampepfl/dotty/issues/13272. For now let's
disable scaladoc for scala 3.

JIRA Issues: CSL-11235

Differential Revision: https://phabricator.twitter.biz/D720887