util/util-slf4j-api: scala3 migration
**Problem**
We're migrating util-slf4j-api to scala3. Dotty has new syntax rules that need
to be addressed. This project also depends on util-mock and therefore on
mockito-scala, which is incompatible with scala3.

**Solution**
Use migration flags to solve syntax warnings. Remove dependency on mockito-scala
and use mockito-core instead in LoggingTest and MarkerLoggingTest.

JIRA Issues: CSL-11110

Differential Revision: https://phabricator.twitter.biz/D725491