util/util-jvm: scala 3 migration
**Problem**
Build util-jvm with Scala 3

**Solution**
# Change util-jvm project settings.
# Compile project + test with Scala 3 migration flags
# Remove flags
# Fix incompatibilities
# Run cross-building tests

**Result**
//Auto-rewrite fixes://
Scala 3 requires parentheses around the parameter of a lambda:
- Jvm.scala
- JvmStatsTest.scala

//Other fixes://
DurationOps expects a Long for as an input for operations.  The conversion to
long used to happen implicitly, but now needs to be explicit in order to work.
- CpuProfile.scala
- CpuProfileTest.scala

Implicit conversion `long2float` in object Long is deprecated,
need to be explicit:
- JvmStats.scala

getter `Stream` is deprecated, use `compat.LazyList` instead:
- CpuProfileTest.scala

Widening conversion from Long to Double is deprecated:
- EstimatorApp.scala

Scala 3 infers that executor is of type `ScheduledExecutorService` when it
should be of type `MockScheduledExecutorService`.  Explicitly give type to avoid
error:
- JvmTest.scala

JIRA Issues: CSL-11094

Differential Revision: https://phabricator.twitter.biz/D728602