util-stats: propagate CounterishGaugeType
Problem:
Regardess of having set `.withCounterishGauge()`,
when building a gauge via the `MetricBuilder.gauge()`
api, we always return a gauge schema. This makes the
`.withCounterishGauge()`, in effect, non-functional.

Solution:
Rather than copying a `MetricBuilder` and resetting
its `metricType`, simply verify that we're only calling
the `.gauge()` `.counter()` or `.stat()` api on a `MetricBuilder`
with the appropriate `metricType`

JIRA Issues: CSL-11280, CSL-11336

Differential Revision: https://phabricator.twitter.biz/D749274