util/util-core: Fix conversions to units from Int in Scala3
This fixes lost source-compatibility for unit conversions from Int values in
Scala3. As discussed in #295 I extracted/adapted this from #290 (previous
discussion).

Previously something like 1.toInt.second did compile in Scala2 (with default
settings) but didn't in Scala3.

I added forwarders for all units which use Long as base, as that seemed
reasonable and safe. I didn't add an implicit conversion from Float to Double
because floating-point types are dumb and I don't trust the conversion to be
lossless in all cases.

Signed-off-by: Yufan Gong <yufang@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D759923