util-jackson: Block unsafe polymorphic base type deserialization
Problem:
Jackson ObjectMapper can suffer from CVEs that exploit the permissive nature of
class-name-based polymorphic deserialization. This is a concern if JSON is
accepted by untrusted clients and polymorphic handling is enabled. The two ways
of enabling polymorphic handling are `enableDefaultTyping()` and annotating
properties with `@JsonTypeInfo`.

Solution:
Util-Jackson's ScalaObjectMapper is not at risk for this type of CVE because
default typing is not enabled by default. Out of security precaution, enable
BLOCK_UNSAFE_POLYMORPHIC_BASE_TYPES in ScalaObjectMapper.

Result:
Turning on this mapper feature blocks the use of a set of "unsafe" base types.
Even if `enableDefaultTyping` is configured on a custom ScalaObjectMapper,
the mapper will be protected from this CVE.

JIRA Issues: CSL-11410

Differential Revision: https://phabricator.twitter.biz/D780863