util-stats: add testing suite for metric expressions
Problem:
It's currently difficult to test for correct
implementation of metric expressions with a short
development loop (ie without having to deploy and
check the admin endpoint).

Solution:
Build a testing suite that demands you define only
the statsReceiver and the downstream label for
metric labels. This suite will make it easy to
get a full summary of registered expressions,
assert specific expressions are instrumented, and
confirm a set of expected expressions matches
the actual expressions registered on the stats receiver.

JIRA Issues: CSL-11282

Differential Revision: https://phabricator.twitter.biz/D783532