util-core: Return suppressed exception information in Closable.sequence
Problem

Closable.sequence returns the first failed Future[Unit] result of closing the Closables, and in the case of multiple failures, the subsequent exceptions are lost.

Solution

Add the subsequent failures to the list of suppressed exceptions attached to the first failed Future[Unit] result.

Result

Users will be able to see all failures in the case that there are multiple failures in Closable.sequence.

JIRA Issues: CSL-11522

Differential Revision: https://phabricator.twitter.biz/D809561