[util-core][finagle-core][fiber-scheduler] introduce withMaxSyncConcurrency
**Problem**

When a `FuturePool` is redirected to a `ForkingScheduler`, the characteristics of the original thread pool are lost. The synchronous concurrency limit imposed be the number of threads and the queue size aren't used.

**Solution**

Introduce a new method `withMaxSyncConcurrency`, which is used to create a new `ForkingScheduler` wrapper with bounded concurrency and queue.

Differential Revision: https://phabricator.twitter.biz/D791343