util/util-slf4j-api: API updates and deprecation removal
Problem

The current usage pattern that Scala users should prefer using
the `Logging` trait is a bit flawed as it doesn't account for the
very common and correct case of creating a Logger once in a
companion object for use in a companion class. This would mirror
the Java best practice of defining a logger instance as a static
member of a Java class. This is generally preferred as a best practice
because there is typically no reason to instantiate a new logger
instance every time the class it is defined in is instantiated.

Related, since loggers are named after the class passed to the
constructor, the API doesn't currently attempt to normalize the
name of the logger when given a class that represents a Scala singleton
object (see: https://docs.scala-lang.org/tour/singleton-objects.html)
meaning that a logger created in a companion object can end up with
the trailing `$` symbol which Scala uses for singleton object class names.

Also, the Logger functions were not optimized in the same way as the
Logging trait (exposing call-by-name versions) since Scala users were
meant to prefer the Logging trait which did have these variations whereas
Java users could prefer the Logger functions directly. These variations
should also exist on the Logger for Scala users.

Additionally, `c.t.util.logging.Loggers` was deprecated some
time ago as it is no longer necessary for bridging the Scala object
constructors for Java users. So let's finally remove it.

Solution

Add `call-by-name` method variations to the Logger to provide the
same behavior available in the Logging trait.

Normalize a logger name when a Scala singleton object class is passed
to drop the `$` suffix.

Remove the deprecated `Loggers` class as Java users can use the
apply methods in `Logger` now directly without additional ceremony.

Result

More utility for Scala users for the Logger. Logger names will be more
consistent across a singleton object or a class. And reduction of tech
debt through the removal of deprecated code.

Differential Revision: https://phabricator.twitter.biz/D820252