util-core: Fix race condition in Var.collectIndependent
Problem

There is a subtle race condition in the current Var.collectIndepdent implementation where the "async" update may be racing
with the very last "sync" one and triggering the observation with "null" in the last position of the collected result.

Solution

Fix race condition by checking the `filling` state within a synchronized block.

JIRA Issues: CSL-11377

Differential Revision: https://phabricator.twitter.biz/D824055