util-slf4j-bridge: Add Slf4jBridge trait for handler installation in class construction
Problem

Installation of the SLF4JBridgeHandler should happen as early as possible, ideally during
the construction of the main application.

Solution

Add a mixin which can be composed into extensions of `c.t.app.App` which will attempt
installation of the bridge handler via the `Slf4jBridgeUtility` as part of class
construction. Move TwitterServer to use the mixin in util-slf4j-jul-bridge.

Result

A simple way for extensions of `c.t.app.App` to ensure installation of the SLF4JBridgeHandler
is attempted during the constructor of the `App`.

Differential Revision: https://phabricator.twitter.biz/D827493