[util-core] lightweight Local getter
**Problem**

In some contexts, `Local`s are read by the same thread at a high rate. For example, a scheduler worker that is pinned to a single thread might need to access a `Local` every time that a task is submitted to it. Even though the `Local` read is done always by the same thread, the `Local` implementation stills need to go through the `ThreadLocal` map to find the local value.

**Solution**

Change the encoding of the `ThreadLocal` that holds the `Local.Context` to use a `ContextRef`. When the context needs to be updated, instead of setting the `ThreadLocal` value, the `ContextRef` object is updated. This is safe because `ContextRef` is always accessed by the owning thread. The `ContextRef` value is used to implement a getter that caches the value of the local. On each getter invocation, the implementation checks if the context hasn't changed and uses the cached value if possible, which also avoids the `Context.get` lookup.

**Notes**

- For performance reasons, the getter implementation doesn't check if it's being accessed by the owning thread. There's an `assert` to make it easier to identify issues in tests.
- I'm going to create a separate review to use this new mechanism in the fiber scheduler implementation
- It seems that this mechanism could be used to implement the cpu time tracing in the `LocalScheduler` efficiently.

**JMH results**

```
[info] Benchmark                      Mode  Cnt  Score   Error  Units
[info] LocalBenchmark.get             avgt   10  4.196 ± 0.213  ns/op
[info] LocalBenchmark.getThreadLocal  avgt   10  1.789 ± 0.104  ns/op
```

Differential Revision: https://phabricator.twitter.biz/D820768