finagle-stats-core: Export counters in Prometheus format
Problem:
We want to export Finagle counters in Prometheus format. This format allows
Finagle to have rich (dimensional) metrics where metrics have labels attached
to them.

Solution:
Encode counters name, value and their labels in the Prometheus format.
Optionally write metadata of the type of metric and the units that the metric
counts in. Write the metric to a StringBuilder.

JIRA Issues: CSL-11562

Differential Revision: https://phabricator.twitter.biz/D839136