util-stats: Some improvements around MetricBuilder to make it more java friendly
Problem

The MetricBuilder api was designed around scala where the named and defaulted
arguments in the `apply` constructor function are nice to use. However, those niceties
don't work in java.

Solution

We already have a bunch of `.withX` methods that form a builder pattern on
MetricBuilder. To make them useful for java folks we add a simpler constructor
function. We also add some aliases for the `MetricType` family of types that are
easier to get at from java.

We also move the compile tests to a package that is outside of `c.t.finagle.stats`
so we can be sure that the visibility of classes is usable from java code that may
live in other packages.

Differential Revision: https://phabricator.twitter.biz/D845297