util-stats: Add builder pattern for labels
Problem

We have labels on the MetricBuilder but we can't add them outside the
class itself outside of the mega apply method.

Solution

Make a builder method to swap out the labels. We also make the labels
private[finagle] while the API is in flux so as to not close doors on
ourselves.

Differential Revision: https://phabricator.twitter.biz/D847815