util-stats: Lower the abstraction level of stats name translation
Problem

We want to be able to scope stats collectively and we normally do that
with NameTranslatingStatsReceiver, but that abstraction only works on
the hierarchical name and doesn't allow us to auto-magically scope
things based on labels.

Solution

Introduce an abstraction just a bit lower, TranslatingStatsReceiver,
which operates on the whole MetricBuilder. This abstraction makes the
sky the limit for modifying metric definitions.

Differential Revision: https://phabricator.twitter.biz/D848642