util/util-core: promise cpu usage resource tracing
Problem:
We want low-level resource information per request.

Solution:
In our Promise implementation, all continuations stem from the internal
`Transformer` and `Monitored` api which are both continuations represented
by `K`.  We have a mutable Local that can keep track of resource usage as each
continuation is run. The local is set in Finagle if the request is actively tracing.
When the continuation is created, we check this local to determine if the
continuation should be tracked

Result:
Read resource data from the local context when tracing a request in finagle.
This includes:
   - Accumulated Cpu Time
   - # of executed continuations

JIRA Issues: CSL-11501, CSL-11627

Differential Revision: https://phabricator.twitter.biz/D791323