util-stats: Remove metric instantiation methods from `MetricBuilder`
Problem

We have a StatsReceiver that lives on MetricBuilder. It doesn't need
to be there. It's currently used mostly for supporting expressions and
their convenient `.build()` method which hunts for a stats receiver in
the MetricSchema to register the expression with.

Solution

In order to unwind this we first need to remove the metric creation methods
from MetricBuilder. In the future we'll make expressions register themselves
with a global instance of a StatsReceiver or even a new interface that
feels more purpose driven.

Differential Revision: https://phabricator.twitter.biz/D859036