util-stats: introduce APIs to create Expressions with StatsReceiver
Problem
When customers use `StringExpression` to create an `Expression`
with a list of metric names, they might need to get the scope of
the `StatsReceiver` they are using, and remove/append metric
names to it in order to match the counter or stats namespace.

Solution
Introduce APIs to create `Expression`s with a `StatsReceiver`.

JIRA Issues: CSL-11827

Differential Revision: https://phabricator.twitter.biz/D861551