util-stats: Add a TranslatingStatsReceiver that will add a label and also prepend the label value as a scope
Problem

We want to be able to define metrics that can be identified by both a hierarchical name
and a dimensional name, but are otherwise the same metric. However, we normally use
"scoped" StatsReceivers which are fundamentally hierarchical only.

Solution

Add a new way of scoping a StatsReceiver that will allow us to add both the hierarchical
prefix and a label at the same time. This will let us progressively migrate our stats by
changing the scoping of StatsReceivers.

JIRA Issues: CSL-11565

Differential Revision: https://phabricator.twitter.biz/D864997