util: Remove the StatsReceiver field from MetricBuilder
Problem

MetricBuilder is a relatively nice representation of a metric definition with
one exception: the statsReceiver field, which defines some behavior and isn't
always possible to populate. This is primarily used for registering expressions
but that is kind of fraught with danger: which `statsReceiver` do you use when
registering an expression composed of multiple metrics?

Solution

Remove that field and instead use explicit registration of expressions with a
StatsReceiver. This pattern is still not perfect but perhaps a step in the right
direction and we can try to sort out how to get rid of this registration pain
later.

JIRA Issues: CSL-11900

Differential Revision: https://phabricator.twitter.biz/D871577