util-stats: Fix some .toString implementations in TranslatingStatsReceivers
Problem

Some customers use the StatsReceiver.toString method to figure out the
scopes of the stats receiver. While an anti-pattern, we need to support
it until we get a better solution for them. However, this only applies
to hierarchical translation but we're doing it for dimensional only
transformations as well.

Solution

Fix these situations and add some tests.

Differential Revision: https://phabricator.twitter.biz/D898036