util-jackson: Use field value type as field type for optional fields
Problem
When deserializing a case class, we interpret the field
type with the field constructor type. For optional field,
Jackson returns `Option[Object]` as the constructor
type. However, field validations would require a specific
field type in order to run the constraint validator. This
causes an exception to throw when an optional field is
validated via deserializing with Jackson.

Solution
Use the filed value type as the field type for optional field.

Differential Revision: https://phabricator.twitter.biz/D915377