util/util-app - Fix Flaggable InetSocketAddress test
Problem

The Flaggable InetSocketAddress test uses a real IP address
which when parsed into an InetSocketAddress and `getHostName` is
called results in a hostname lookup and SecurityManager check
to see if a connection to that host is allowed. For unit
test meant to run in CI the installed SecurityManager fails
this check since connections off of localhost is (rightly) not
allowed.

Solution

Don't use an external IP address. Use the loopback, `127.0.0.1`
which should allow to test the same functionality without
violating the SecurityManager. Additionally fix the remote
getHostname asseertion. Note this is assuming that the machine
`/etc/hosts` file maps `localhost` to the loopback `127.0.0.1`
which should be valid in almost all cases but we may hit a
case where for some reason this is not true. But for now this
should work as expected.

JIRA Issues: CSL-12163

Differential Revision: https://phabricator.twitter.biz/D929587