finagle: Introduce configurable summary for histograms
Problem

We want to allow the developers to decide the verbosity of the histograms that they
create with`StatsReceiver`.

Solution

Introduce a new param for `MetricsBuilder` to override the default format for
histograms (`HistogramFormat`). A few formats are already supported, including
`NoSummary` and `ShortSummary`, both of which will reduce the histogram
footprint.

The exporting subsystem (TwitterServer) already respects the new parameter.

The default histogram format remains unchanged and is the same as `FullSummary`.
That said, this commit doesn't introduce a runtime change.

JIRA Issues: CSL-12237

Differential Revision: https://phabricator.twitter.biz/D908189